Enabling img_sigs and add a toggle on the main screen
Have currently introduced most of the popular Linux images
and utilities for signature checking, if things go smoothly,
I'll continue to add addtional checks for the other images.

I'm currently avoiding development and rolling releases as the
files are typically not versioned and can introduce a lot of breakage
as the files can be replaced by build procedures and are difficult to
track.

If there are any problems, please open up an issue.