Ensure proper release is checked out
CI may have been checking out development even though it was
doing everything else for release branch so when tags were applied
it was applied on the wrong branch.

Development and master should have been near the same except for
release bump.

Hard setting the checkout to the proper branch should fix this.

Closes: https://github.com/netbootxyz/netboot.xyz/issues/1294