fix(ipxe): fix filename checks
The `filename` contains the full path to the file, usually the netboot
assets are in `/` of a tftp server, so the check if its booted from the
ipxe assets will always fail. Add an additional check prepended with `/`
to allow the check to pass when booted from the root of a tftp server.

Also add loading of MAC- ipxe scripts with hypens in addition to raw, i.e.
`MAC-00-11-22-33-44-55.ipxe` or `MAC-001122334455.ipxe`.

Fixes: #1097